+++ Another-Mirai-Native/Adapter/MiraiEventArgs/BotOfflineEventActive.cs
@@ -9,6 +9,6 @@ namespace Another_Mirai_Native.Adapter.MiraiEventArgs
     public class BotOfflineEventActive
     {
         public string type { get; set; }
-        public int qq { get; set; }
+        public long qq { get; set; }
     }
 }
