+++ Another-Mirai-Native/Adapter/MiraiEventArgs/BotOnlineEvent.cs
@@ -9,6 +9,6 @@ namespace Another_Mirai_Native.Adapter.MiraiEventArgs
     public class BotOnlineEvent
     {
         public string type { get; set; }
-        public int qq { get; set; }
+        public long qq { get; set; }
     }
 }
