+++ Another-Mirai-Native/Adapter/MiraiEventArgs/FriendRecallEvent.cs
@@ -9,9 +9,9 @@ namespace Another_Mirai_Native.Adapter.MiraiEventArgs
     public class FriendRecallEvent
     {
         public string type { get; set; }
-        public int authorId { get; set; }
-        public int messageId { get; set; }
-        public int time { get; set; }
-        public int _operator { get; set; }
+        public long authorId { get; set; }
+        public long messageId { get; set; }
+        public long time { get; set; }
+        public long _operator { get; set; }
     }
 }
