+++ Another-Mirai-Native/Adapter/MiraiEventArgs/NudgeEvent.cs
@@ -9,14 +9,14 @@ namespace Another_Mirai_Native.Adapter.MiraiEventArgs
     public class NudgeEvent
     {
         public string type { get; set; }
-        public int fromId { get; set; }
+        public long fromId { get; set; }
         public Subject subject { get; set; }
         public string action { get; set; }
         public string suffix { get; set; }
-        public int target { get; set; }
+        public long target { get; set; }
         public class Subject
         {
-            public int id { get; set; }
+            public long id { get; set; }
             public string kind { get; set; }
         }
