+++ Another-Mirai-Native/Adapter/MiraiMessageEventArgs/GroupMessage.cs
@@ -17,14 +17,14 @@ namespace Another_Mirai_Native.Adapter.MiraiMessageEventArgs
             public string memberName { get; set; }
             public string specialTitle { get; set; }
             public string permission { get; set; }
-            public int joinTimestamp { get; set; }
-            public int lastSpeakTimestamp { get; set; }
-            public int muteTimeRemaining { get; set; }
+            public long joinTimestamp { get; set; }
+            public long lastSpeakTimestamp { get; set; }
+            public long muteTimeRemaining { get; set; }
             public Group group { get; set; }
         }
         public class Group
         {
-            public int id { get; set; }
+            public long id { get; set; }
             public string name { get; set; }
             public string permission { get; set; }
         }
